requirements.txt
# Base
fastapi>=0.104.1
uvicorn==0.22.0
pydantic>=2.5.2
sqlalchemy>=2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
fastapi-users
pydantic[email]==2.10.1
# RSS
feedparser

# Telegram
python-telegram-bot>=20.7

# Discord
discord.py>=2.3.2

# WhatsApp
whatsapp-api-client-python==0.0.47  # Change from >=2.0.0 to ==0.0.47

# Database
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Web
requests==2.31.0
python-multipart>=0.0.6

# Async
aiohttp>=3.9.1
asyncio>=3.4.3

# Extras
python-dotenv==1.0.0
fastapi-mail==1.4.1
email-validator==2.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.24.0
